minor style update moved component
